clan-competition-bot/requirements.txt
discord.py
python-dotenv
watchdog
pymongo
aiohttp